fix example tabs
